fix: separate google sheets utils to allow usage in browser (#716)
