+++ __tests__/api-atlases-id-source-datasets-id.test.ts
@@ -7,7 +7,7 @@ import {
 import { AtlasSourceDatasetEditData } from "../app/apis/catalog/hca-atlas-tracker/common/schema";
 import { METHOD } from "../app/common/entities";
 import { endPgPool, query } from "../app/services/database";
-import { getSheetTitleForApi } from "../app/utils/google-sheets";
+import { getSheetTitleForApi } from "../app/utils/google-sheets-api";
 import sourceDatasetHandler from "../pages/api/atlases/[atlasId]/source-datasets/[sourceDatasetId]";
 import {
   ATLAS_WITH_MISC_SOURCE_STUDIES,
@@ -51,13 +51,12 @@ jest.mock("next-auth");
 
 const getSheetTitleMock = getSheetTitleForApi as jest.Mock;
 
-jest.mock("../app/utils/google-sheets", () => {
-  const googleSheets: typeof import("../app/utils/google-sheets") =
-    jest.requireActual("../app/utils/google-sheets");
+jest.mock("../app/utils/google-sheets-api", () => {
+  const googleSheetsApi: typeof import("../app/utils/google-sheets-api") =
+    jest.requireActual("../app/utils/google-sheets-api");
 
   return {
-    InvalidSheetError: googleSheets.InvalidSheetError,
-    getSheetTitleForApi: jest.fn(googleSheets.getSheetTitleForApi),
+    getSheetTitleForApi: jest.fn(googleSheetsApi.getSheetTitleForApi),
   };
 });
